fix: header band height and add useNavigation to get navigation object
