+++ packages/core/src/header/HeaderBandProvider.ts
@@ -1,13 +1,6 @@
-import {
-  useCallback,
-  useEffect,
-  useLayoutEffect,
-  useMemo,
-  useState,
-} from 'react';
 import {HeaderBandItem} from './types';
 
-const HEADER_DEFAULT_HEIGHT = 80;
+const HEADER_DEFAULT_HEIGHT = 88;
 const HEADER_BAND_HEIGHT = 24;
 
 class HeaderBandProvider {
@@ -88,52 +81,3 @@ class HeaderBandProvider {
 }
 
 export const headerBandProvider = new HeaderBandProvider();
-
-export const useHeaderBands = (navigation: any) => {
-  const [state, setState] = useState<{
-    bands: HeaderBandItem[];
-    headerHeight: number;
-  }>({
-    bands: headerBandProvider.getFilteredBands(),
-    headerHeight: headerBandProvider.getheaderHeight(),
-  });
-
-  useEffect(() => {
-    headerBandProvider.register(setState);
-  }, []);
-
-  const _addHeaderBand = useCallback(
-    (band: HeaderBandItem) => headerBandProvider.addHeaderBand(band),
-    [],
-  );
-
-  const _updateHeaderBand = useCallback(
-    (key: string, newBand: HeaderBandItem) =>
-      headerBandProvider.updateHeaderBand(key, newBand),
-    [],
-  );
-
-  const _removeHeaderBand = useCallback(
-    (key: string) => headerBandProvider.removeHeaderBand(key),
-    [],
-  );
-
-  useLayoutEffect(() => {
-    if (navigation) {
-      navigation.setOptions({
-        headerStyle: {
-          height: state.headerHeight,
-        },
-      });
-    }
-  }, [navigation, state.headerHeight]);
-
-  return useMemo(() => {
-    return {
-      ...state,
-      addHeaderBand: _addHeaderBand,
-      updateHeaderBand: _updateHeaderBand,
-      removeHeaderBand: _removeHeaderBand,
-    };
-  }, [state, _addHeaderBand, _updateHeaderBand, _removeHeaderBand]);
-};
