+++ packages/core/src/header/index.ts
@@ -17,7 +17,7 @@
  */
 
 export {headerActionsProvider, useHeaderOptions} from './HeaderActionsProvider';
-export {headerBandProvider, useHeaderBands} from './HeaderBandProvider';
+export {headerBandProvider} from './HeaderBandProvider';
 export {useBasicActions} from './hooks';
 export {
   HeaderActions,
@@ -25,3 +25,4 @@ export {
   ActionType,
   HeaderBandItem,
 } from './types';
+export {useHeaderBands} from './use-header-band';
